% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_gsva.R
\name{cgsva}
\alias{cgsva}
\title{Gene Set Variation Analysis for cytokines (cGSVA)}
\usage{
cgsva(eset, design, db)
}
\arguments{
\item{eset}{}

\item{design}{}

\item{db}{}
}
\description{
Gene Set Variation Analysis for cytokines (cGSVA)
}
