% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_cgsvar.R
\name{cgsvar}
\alias{cgsvar}
\title{Calculate coefficients from variable selection using GSVA and PLS-DA}
\usage{
cgsvar(eset, y, obs_id, db)
}
\arguments{
\item{eset}{Expression Set object containing gene expression data.}

\item{y}{Response variable}

\item{obs_id}{Observation ID or sample if looking there are biological replicates}

\item{db}{ligand-receptor database}
}
\value{
A named vector of ligands indicating importance
}
\description{
Uses a multiple regression approach when computing receptor weights with
Partial Least Squares - Discriminant Analysis and pools these receptors
together to compute a value for the ligand or cytokine given a gene set.
}
