% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_cpca.R
\name{cpca}
\alias{cpca}
\title{Calculate top ligands using a PCA approach and use for weights to}
\usage{
cpca(eset, y, obs_id, db)
}
\arguments{
\item{eset}{Expression Set object containing gene expression data.}

\item{y}{}

\item{obs_id}{}

\item{db}{}
}
\value{
List of differentially expressed ligands ordered by p-values
}
\description{
Calculate top ligands using a PCA approach and use for weights to
}
