% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_cpca_plsda.R
\name{cpca_plsda}
\alias{cpca_plsda}
\title{Title}
\usage{
cpca_plsda(eset, treatment, db)
}
\arguments{
\item{eset}{Expression Set object containing gene expression data.}

\item{treatment}{Treatment response variable}

\item{db}{ligand-receptor database}
}
\description{
Title
}
