% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_run_cytokinefinder.R
\name{cytokinefinder}
\alias{cytokinefinder}
\title{Title}
\usage{
cytokinefinder(eset, design, dbs, methods)
}
\arguments{
\item{eset}{}

\item{design}{}

\item{dbs}{}

\item{methods}{}
}
\description{
Title
}
