% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\docType{data}
\name{cytosig_beta}
\alias{cytosig_beta}
\title{CytoSig Betas}
\format{
A matrix of cytosig gene co-expression weights or betas
}
\source{
From CytoSig
}
\usage{
cytosig_beta
}
\description{
CytoSig Betas
}
\keyword{datasets}
