% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_cytosig_custom_ridge.R
\name{cytosig_custom_ridge}
\alias{cytosig_custom_ridge}
\title{Run CytoSig Custom Ridge Regression function}
\usage{
cytosig_custom_ridge(
  eset,
  design,
  obs_id = NULL,
  correlation = NULL,
  beta_coef = cytosig_beta
)
}
\arguments{
\item{eset}{Expression matrix of microarray or RNAseq experiments}

\item{design}{design matrix list that indicates conditions to compare and if paired design}

\item{obs_id}{if paired design, indicate observation ID to map biological replicates to sample of origin if paired}

\item{correlation}{if paired design, indicate correlation blocks for paired data}

\item{beta_coef}{beta matrix for ridge regression. Defaults to cytosig_beta included in package.}
}
\value{
CytoSig ridge regression results table
}
\description{
Run CytoSig Custom Ridge Regression function
}
\details{
Cytokine Signaling Analyzer or CytoSig (Jiang et al, 2021) is a ridge regression predictive model of cytokine signaling
cascades trained on 20,591 transcriptomic profiles of perturbations relevant to cytokine activity. The following is our
adoption of CytoSig using the beta coefficients published using logFC results generated differential expression through
our run_limma function based on the standard of practice, which can be done through cytokineFindeR's functions.
}
\examples{
set.seed(42)
genes   <- paste0("GENE", 1:30)
samples <- paste0("S", 1:8)
eset    <- matrix(rnorm(240), nrow = 30, ncol = 8,
                 dimnames = list(genes, samples))
treatment <- rep(c("ctrl", "trt"), each = 4)
design  <- model.matrix(~ treatment)
rownames(design) <- samples
# synthetic beta coefficient matrix with matching gene names
beta    <- matrix(rnorm(60), nrow = 30, ncol = 2,
                 dimnames = list(genes, c("CytokineA", "CytokineB")))
result  <- cytosig_custom_ridge(eset, design, beta_coef = beta)

}
