% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_cpca_plsda.R
\docType{data}
\name{db}
\alias{db}
\title{Title}
\format{
An object of class \code{list} of length 2293.
}
\usage{
db
}
\arguments{
\item{eset}{Expression Set object containing gene expression data.}

\item{treatment}{Treatment response variable}

\item{db}{ligand-receptor database}
}
\description{
Title
}
\keyword{datasets}
