% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\docType{data}
\name{dbs_all}
\alias{dbs_all}
\title{Database List unfiltered}
\format{
A list of lists containing LRIs for all database sources
}
\source{
Various ligand receptor interaction databases
}
\usage{
dbs_all
}
\description{
Database List unfiltered
}
\keyword{datasets}
