% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\docType{data}
\name{dbs_cytosig}
\alias{dbs_cytosig}
\title{Database List to make comparable benchmarks against CytoSig}
\format{
A list of lists containing LRIs for all database sources subset on CytoSig features
}
\source{
Various LRI databases
}
\usage{
dbs_cytosig
}
\description{
Database List to make comparable benchmarks against CytoSig
}
\keyword{datasets}
