% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\docType{data}
\name{dbs_subset}
\alias{dbs_subset}
\title{Small Database example}
\format{
A list of lists containing LRIs for a subset of database sources
}
\source{
Various ligand receptor interaction databases
}
\usage{
dbs_subset
}
\description{
Small Database example
}
\keyword{datasets}
