% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download_from_url.R
\name{download_from_url}
\alias{download_from_url}
\title{Download a file from a URL if it doesn't exist locally.
This function checks if the specified destination file exists locally.
If not, it downloads the file from the provided URL to the destination file.
Optionally, it can extract the contents if the file is a zip archive.}
\usage{
download_from_url(
  url,
  destFile,
  mode = NULL,
  extractZip = FALSE,
  extractDir = NULL
)
}
\arguments{
\item{url}{The URL of the file to download.}

\item{destFile}{Destination file path where the downloaded file should be saved.}

\item{mode}{Optional. Mode for downloading the file. Default is NULL (automatic detection).
Use "wb" for binary files like zip archives.}

\item{extractZip}{Logical, whether to extract the file if it's a zip archive (default is FALSE).}

\item{extractDir}{Optional. Directory where the contents of the zip file should be extracted.
If NULL, extracts to the current working directory.}
}
\description{
Download a file from a URL if it doesn't exist locally.
This function checks if the specified destination file exists locally.
If not, it downloads the file from the provided URL to the destination file.
Optionally, it can extract the contents if the file is a zip archive.
}
