% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_extract_db.R
\name{extract_from_db}
\alias{extract_from_db}
\title{Helper function to filter database and extract ligands-receptors pairs match}
\usage{
extract_from_db(cytokine = NULL, eset, dbs)
}
\arguments{
\item{cytokine}{An optional vector of cytokines that can be added to address
some specifically named receptors that may have been omitted due to character
changes such as TNFA vs TNF.}

\item{eset}{}

\item{dbs}{}
}
\value{
A list of lists showing a set of ligand genes where each ligand
contains a list of genes that were subset defined as "receptor" genes
}
\description{
Provide a list of databases and match ligands of interest.
}
\details{
Iterate through the expression matrix input and match a vector of genes
of interest.
}
