% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_eset_and_db.R, R/utils_extract_db.R
\name{filter_db_against_eset}
\alias{filter_db_against_eset}
\title{Filter database and extract ligands-receptors pairs that match the eset data}
\usage{
filter_db_against_eset(eset, dbs)

filter_db_against_eset(eset, dbs)
}
\arguments{
\item{eset}{A numeric matrix that represents the expression set}

\item{dbs}{A nested list of lists containing the databases, the ligand genes,
and the receptor genes for each ligand gene}
}
\value{
A list of lists showing a set of ligand genes where each ligand
contains a list of genes that were subset defined as "receptor" genes

A list of lists showing a set of ligand genes where each ligand
contains a list of genes that were subset defined as "receptor" genes
}
\description{
Provide a list of databases and match against receptors of interest.
Remove ligands that do not have

Provide a list of databases and match against receptors of interest.
Remove ligands that do not have
}
\details{
Iterate through the expression matrix input and match a vector of genes
of interest.

Iterate through the expression matrix input and match a vector of genes
of interest.
}
