% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ligand_summary}
\alias{plot_ligand_summary}
\title{Plot Ligand of Interest summary result across methods and databases}
\usage{
plot_ligand_summary(data)
}
\arguments{
\item{data}{}
}
\description{
Plot Ligand of Interest summary result across methods and databases
}
