% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ligands.R
\name{process_method_db}
\alias{process_method_db}
\title{Define a function to process each method and database}
\usage{
process_method_db(df, method_name, db_name)
}
\arguments{
\item{df}{}

\item{method_name}{}

\item{db_name}{}
}
\description{
Define a function to process each method and database
}
