% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_eset_and_db.R
\name{remove_zero_variance_ligands}
\alias{remove_zero_variance_ligands}
\title{Title}
\usage{
remove_zero_variance_ligands(eset, dbs)
}
\arguments{
\item{eset}{}

\item{dbs}{}
}
\description{
Title
}
