% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ligands.R
\name{reshape_metric}
\alias{reshape_metric}
\title{Secondary helper function to reshape a specific metric into the desired format}
\usage{
reshape_metric(df, metric, type)
}
\arguments{
\item{df}{}

\item{metric}{}

\item{type}{}
}
\description{
Secondary helper function to reshape a specific metric into the desired format
}
