% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_save_data_to_rds.R
\name{save_in_batch}
\alias{save_in_batch}
\title{Save multiple GEO datasets in batch}
\usage{
save_in_batch(geo_ids, dir_path = "data/")
}
\arguments{
\item{geo_ids}{A vector of character strings representing the GEO IDs}

\item{dir_path}{A character string representing the directory path 
where the RDS files should be saved. DEFAULT: data/}
}
\value{
A list of GEO with GEO IDs as names
}
\description{
This function saves multiple GEO datasets to separate RDS files by
calling `save_to_rds` on each GEO ID.
}
