% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_save_data_to_rds.R
\name{save_to_rds_from_geo}
\alias{save_to_rds_from_geo}
\title{Save data to RDS or RDA format}
\usage{
save_to_rds_from_geo(geo_id, dir_path = "data/")
}
\arguments{
\item{geo_id}{A character string representing the GEO ID
to be downloaded}

\item{dir_path}{A character string representing the directory path
where the RDS files should be saved. DEFAULT: data/}
}
\value{
The GEO data if successful; otherwise, return NULL
}
\description{
This function fetches GEO data using the geoQuery package
and saves it as an RDS
}
