% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ligands.R
\name{summarize_df}
\alias{summarize_df}
\title{Primary helper function to create the summary DataFrame}
\usage{
summarize_df(results_df)
}
\arguments{
\item{results_df}{}
}
\description{
Primary helper function to create the summary DataFrame
}
