% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_cfgsea.R
\name{cfgsea}
\alias{cfgsea}
\title{Title}
\usage{
cfgsea(eset, design, db)
}
\arguments{
\item{eset}{Expression Set object containing gene expression data.}

\item{design}{Design matrix generated from create_design()}

\item{db}{Ligand-receptor database}
}
\value{
a table with GSEA results. Each row corresponds to a ligand
}
\description{
Title
}
