% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_cfgsea.R
\name{cfgsea}
\alias{cfgsea}
\title{Gene Set Enrichment Analysis}
\usage{
cfgsea(eset, design, db, obs_id = NULL, correlation = NULL)
}
\arguments{
\item{eset}{Expression Set object containing gene expression data.}

\item{design}{Design matrix generated from create_design()}

\item{db}{Ligand-receptor database}
}
\value{
a table with GSEA results. Each row corresponds to a ligand
}
\description{
Gene Set Enrichment Analysis
}
