% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_gsva.R
\name{cgsva}
\alias{cgsva}
\title{Gene Set Variation Analysis for cytokines (cGSVA)}
\usage{
cgsva(eset, design, db, obs_id = NULL, correlation = NULL)
}
\arguments{
\item{eset}{Expression Set object containing gene expression data.}

\item{design}{Design matrix generated from create_design()}

\item{db}{Ligand-receptor database}

\item{obs_id}{}

\item{correlation}{}

\item{dupCor}{}
}
\description{
Gene Set Variation Analysis for cytokines (cGSVA)
}
