% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clean_data.R
\name{clean_eset}
\alias{clean_eset}
\title{Given an expression set, make sure that data is
cleaned by gene name using a provided gene list data frame}
\usage{
clean_eset(eset, gene_list_df)
}
\arguments{
\item{eset}{Expression Set as a numeric matrix}

\item{gene_list_df}{}
}
\value{
cleaned up expression matrix
}
\description{
Given an expression set, make sure that data is
cleaned by gene name using a provided gene list data frame
}
