% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_cpca.R
\name{cpca}
\alias{cpca}
\title{Calculate top ligands using a PCA approach and use for weights to}
\usage{
cpca(eset, design, db)
}
\arguments{
\item{eset}{Expression Set object containing gene expression data.}

\item{design}{Design matrix generated from create_design()}

\item{db}{Ligand-receptor database}
}
\value{
a data frame of differentially expressed ligands ordered by p-values
}
\description{
Calculate top ligands using a PCA approach and use for weights to
}
