% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_create_design.R
\name{create_design}
\alias{create_design}
\title{Helper fun to create the design matrix}
\usage{
create_design(treatment, obs_id = NULL, eset = NULL)
}
\arguments{
\item{treatment}{Treatment response variable}

\item{obs_id}{Observation ID: some samples may have unique IDs but come from
the same tissue of origin, if that exists, provide a vector of this to make
sure the Expression matrix accounts for this to avoid incorrect DEA input.}

\item{eset}{Expression matrix (genes x samples); required when \code{obs_id}
is provided in order to compute duplicate correlations via
\code{limma::duplicateCorrelation}.}
\value{
Either a design matrix or a design list with components
}
\description{
Helper fun to create the design matrix
}
\examples{
\dontrun{
# data(golimumab)
# For paired datasets (biological replicates), we can use dupcor blocks
# design <- create_design(treatment = golimumab$cond, obs_id = golimumab$obs_id, eset = golimumab$qc_eset)

# If unpaired datasets, we can ignore the obs_id argument
# design <- create_design(treatment = golimumab$cond, eset = golimumab$qc_eset)
}
}
