% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_create_design.R,
%   R/utils_create_design_mat.R
\name{create_design}
\alias{create_design}
\title{Helper fun to create the design matrix}
\usage{
create_design(y, obs_id)

create_design(y, obs_id)
}
\arguments{
\item{y}{}

\item{obs_id}{}
}
\description{
Helper fun to create the design matrix

Create design matrix for the data based on the outcome and the samples
}
