% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_run_cytokinefinder.R
\name{cytokinefinder}
\alias{cytokinefinder}
\title{The wrapper function to run the benchmarking}
\usage{
cytokinefinder(eset, design, dbs, methods, treatment = NULL)
}
\arguments{
\item{eset}{An expression set (numeric matrix) of genes x samples}

\item{design}{The design matrix for the data set used to generate the model}

\item{dbs}{The databases in a list of list format (for package default, use dbs_all)}

\item{methods}{A vector of methods contained in this package}

\item{treatment}{A vector containing the treatment (specific to the demo
data set, this is to analyze differentially expressed genes between week 0
and week 6) with the drug gollimumab on Ulcerative colitis patients)}
}
\value{
A large BenchmarkResults object containing a nested list of methods
and the results
}
\description{
The wrapper function to run the benchmarking
}
