% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_run_cytokinefinder.R
\name{cytokinefinder}
\alias{cytokinefinder}
\title{The wrapper function to run the benchmarking}
\usage{
cytokinefinder(eset, design, dbs, methods)
}
\arguments{
\item{eset}{}

\item{design}{}

\item{dbs}{}

\item{methods}{}
}
\value{
A large BenchmarkResults object containing a nested list of methods
and the results
}
\description{
The wrapper function to run the benchmarking
}
