% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_cytosig_custom_ridge.R
\name{cytosig_custom_ridge}
\alias{cytosig_custom_ridge}
\title{Run CytoSig Custom Ridge Regression function}
\usage{
cytosig_custom_ridge(
  eset,
  design,
  obs_id = NULL,
  correlation = NULL,
  beta_coef
)
}
\arguments{
\item{eset}{Expression matrix of microarray or RNAseq experiments}

\item{design}{design matrix list that indicates conditions to compare and if paired design}

\item{obs_id}{if paired design, indicate observation ID to map biological replicates to sample of origin if paired}

\item{correlation}{if paired design, indicate correlation blocks for paired data}

\item{beta_coef}{beta matrix for ridge regression from published CytoSig}
}
\value{
CytoSig ridge regression results table
}
\description{
Run CytoSig Custom Ridge Regression function
}
