% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golimumab.R
\docType{data}
\name{golimumab}
\alias{golimumab}
\title{Golimumab Dataset}
\format{
A data object with metadata and expression set of GSE92415 specific to golimumab treatment:
\describe{
\item{variable1}{Description of variable1}
\item{variable2}{Description of variable2}
...
}
}
\usage{
data(golimumab)
}
\description{
Description of the dataset and its variables.
}
\keyword{datasets}
