% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_gsva_limma.R
\name{gsva_limma}
\alias{gsva_limma}
\title{Gene Set Variation Analysis for cytokines (cGSVA)}
\usage{
gsva_limma(eset, design, db, obs_id = NULL, correlation = NULL)
}
\arguments{
\item{eset}{Expression Set object containing gene expression data.}

\item{design}{Design matrix generated from create_design()}

\item{db}{Ligand-receptor database}

\item{obs_id}{Optional: provide a vector of sample IDs making sure the order matches with the eset}

\item{correlation}{Optional: input the correlation consensus between the samples to evaluate if it is paired data}
}
\description{
Gene Set Variation Analysis for cytokines (cGSVA)
}
