% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_gsva_plsda.R
\name{gsva_plsda}
\alias{gsva_plsda}
\title{Calculate coefficients from variable selection using GSVA and PLS-DA}
\usage{
gsva_plsda(eset, treatment, db)
}
\arguments{
\item{eset}{Expression Set object containing gene expression data.}

\item{treatment}{Treatment response variable}

\item{db}{ligand-receptor database}
}
\value{
A named vector of ligands indicating importance
}
\description{
Uses a classification approach when computing receptor weights with
Partial Least Squares - Discriminant Analysis and pools these receptors
together to compute a value for the ligand or cytokine given a gene set.
}
