% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_cpca_limma.R
\name{pca_limma}
\alias{pca_limma}
\title{Calculate top ligands using a PCA approach from receptor genes given a database}
\usage{
pca_limma(eset, design, db, obs_id = NULL, correlation = NULL)
}
\arguments{
\item{eset}{Expression Set object containing gene expression data.}

\item{design}{Design matrix generated from create_design()}

\item{db}{Ligand-receptor database}

\item{obs_id}{Optional: provide a vector of sample IDs making sure the order matches with the eset}

\item{correlation}{Optional: input the correlation consensus between the samples to evaluate if it is paired data}
}
\value{
a data frame of differentially expressed ligands ordered by p-values
}
\description{
Calculate top ligands using a PCA approach from receptor genes given a database
}
