% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_cpca_plsda.R
\name{pca_plsda}
\alias{pca_plsda}
\title{Perform PCA and fit a multivariate regression to measure ligand activity}
\usage{
pca_plsda(eset, treatment, db)
}
\arguments{
\item{eset}{Expression Set object containing gene expression data.}

\item{treatment}{Treatment response variable}

\item{db}{ligand-receptor database}
}
\description{
Perform PCA and fit a multivariate regression to measure ligand activity
}
\examples{
# Load the database
# dbs_all <- dbs_all
# cpca_plsda(eset, treatment, dbs_all$baderlab)
}
