% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ligand_summary}
\alias{plot_ligand_summary}
\title{Plot Ligand of Interest summary result across methods and databases}
\usage{
plot_ligand_summary(data)
}
\arguments{
\item{data}{}
}
\value{
A bar plot figure showing the different methods+database combinations
}
\description{
Plot Ligand of Interest summary result across methods and databases
}
\examples{
#Example usage:
#Assuming 'res' is the output from extract_ligand
\dontrun{
plot_ligand_summary(res)
}
}
