% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preprocess_eset.R
\name{preprocess_eset}
\alias{preprocess_eset}
\title{Preprocess the eset}
\usage{
preprocess_eset(eset, dbs)
}
\arguments{
\item{eset}{the input eset typically the raw gene expression matrix}

\item{dbs}{the list of databases in the list of lists that can be called from the package's internal data}
}
\value{
The filtered expression set that contains 0 variability ligands
}
\description{
Preprocess the eset
}
\examples{
set.seed(42)
genes <- paste0("GENE", 1:20)
eset  <- matrix(rnorm(160), nrow = 20, ncol = 8,
                dimnames = list(genes, paste0("S", 1:8)))
dbs   <- list(db1 = list(LigandA = genes[1:5], LigandB = genes[6:10]))
result <- preprocess_eset(eset, dbs)
names(result)
}
