% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_eset_and_db.R
\name{preprocess_eset}
\alias{preprocess_eset}
\title{Preprocess the eset}
\usage{
preprocess_eset(eset, dbs)
}
\arguments{
\item{eset}{}

\item{dbs}{}
}
\value{
The filtered expression set that contains 0 variability ligands
}
\description{
Preprocess the eset
}
