% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preprocess_eset.R
\name{remove_zero_variance_ligands}
\alias{remove_zero_variance_ligands}
\title{Remove zero variance ligands from the database list of list}
\usage{
remove_zero_variance_ligands(eset, dbs)
}
\arguments{
\item{eset}{}

\item{dbs}{}
}
\value{
A list of databases with ligands removed if there is 0 variance based on the eset
}
\description{
Remove zero variance ligands from the database list of list
}
