% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_retrieve_geo.R
\name{retrieve_geo}
\alias{retrieve_geo}
\title{Retrieve GEO data using the Bioconductor package GEOquery, clean it up, and store in a df list}
\usage{
retrieve_geo(geo_id)
}
\arguments{
\item{geo_id}{A character string representing the GEO ID (e.g. "GSE92415")}
}
\value{
A combined list of data frames saved by GEO_ID series matrix containing metadata, the eset, and annotations (if available)
}
\description{
Retrieve GEO data using the Bioconductor package GEOquery, clean it up, and store in a df list
}
\examples{
\donttest{
geo_datasets <- retrieve_geo("GSE92415")
geo_datasets$GSE92415_series_matrix.txt.gz$metadata
}

}
