% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clean_data.R, R/utils_retrieve_geo.R
\name{retrieve_geo}
\alias{retrieve_geo}
\title{Retrieve GEO data using the Bioconductor package GEOquery, clean it up, and store in a df list}
\usage{
retrieve_geo(geo_id)

retrieve_geo(geo_id)
}
\arguments{
\item{geo_ids}{A character string representing the GEO ID}
}
\description{
Retrieve GEO data using the Bioconductor package GEOquery, clean it up, and store in a df list

Retrieve GEO data using the Bioconductor package GEOquery, clean it up, and store in a df list
}
