% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_run_limma.R
\name{run_limma}
\alias{run_limma}
\title{Run limma in unpaired or paired mode}
\usage{
run_limma(eset, design, obs_id = NULL, correlation = NULL)
}
\arguments{
\item{eset}{expression matrix}

\item{design}{design matrix using the create_design function which indicates blocks for paired samples}

\item{obs_id}{required to indicate the blocks that map biological replicates to the same sample}

\item{correlation}{the average estimated inter-duplicate correlation. The average is the trimmed mean of the individual correlations on the atanh-transformed scale.}
}
\value{
top table that runs differential gene expression analysis. Main purpose for this is to get logFC of all genes for CytoSig input.
}
\description{
Run limma in unpaired or paired mode
}
