% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lri_methods.R
\name{run_lri_methods}
\alias{run_lri_methods}
\title{The core function to run benchmarking of several methods}
\usage{
run_lri_methods(
  eset,
  design,
  dbs,
  methods,
  treatment = NULL,
  obs_id = NULL,
  correlation = NULL,
  verbose = FALSE
)
}
\arguments{
\item{eset}{An expression set (numeric matrix) of genes x samples}

\item{design}{The design matrix for the data set used to generate the model}

\item{dbs}{The databases in a list of list format (for package default, use dbs_all)}

\item{methods}{A character vector of method names to benchmark. Must be one
or more of: \code{"gsva_limma"}, \code{"pca_limma"}, \code{"cfgsea"},
\code{"gsva_plsda"}, \code{"pca_plsda"}, \code{"cytosig_custom_ridge"}.}

\item{treatment}{A vector containing the treatment (specific to the demo
data set, this is to analyze differentially expressed genes between week 0
and week 6) with the drug gollimumab on Ulcerative colitis patients)}

\item{obs_id}{A vector of sample IDs}

\item{correlation}{the average estimated inter-duplicate correlation for}

\item{verbose}{Logical; if \code{TRUE}, progress messages are printed for each
method-database combination. Default is \code{FALSE}.}
}
\value{
A large BenchmarkResults object containing a nested list of methods
and the results
}
\description{
The core function to run benchmarking of several methods
}
\examples{
set.seed(42)
genes   <- paste0("GENE", 1:50)
samples <- paste0("S", 1:8)
eset    <- matrix(rnorm(400), nrow = 50, ncol = 8,
                 dimnames = list(genes, samples))
treatment <- rep(c("ctrl", "trt"), each = 4)
design  <- model.matrix(~ treatment)
rownames(design) <- samples
dbs     <- list(db1 = list(LigandA = genes[1:10], LigandB = genes[11:20]))
result  <- run_lri_methods(eset, design, dbs, methods = "cfgsea",
                           treatment = treatment)
}
