% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lri_methods.R
\name{run_lri_methods}
\alias{run_lri_methods}
\title{The core function to run benchmarking of several methods}
\usage{
run_lri_methods(
  eset,
  design,
  dbs,
  methods,
  treatment = NULL,
  obs_id = NULL,
  correlation = NULL
)
}
\arguments{
\item{eset}{An expression set (numeric matrix) of genes x samples}

\item{design}{The design matrix for the data set used to generate the model}

\item{dbs}{The databases in a list of list format (for package default, use dbs_all)}

\item{methods}{A vector of methods contained in this package}

\item{treatment}{A vector containing the treatment (specific to the demo
data set, this is to analyze differentially expressed genes between week 0
and week 6) with the drug gollimumab on Ulcerative colitis patients)}

\item{obs_id}{A vector of sample IDs}

\item{correlation}{the average estimated inter-duplicate correlation for}

\item{dupCor}{A logical TRUE/FALSE to determine whether or not . Default is set to FALSE.}
}
\value{
A large BenchmarkResults object containing a nested list of methods
and the results
}
\description{
The core function to run benchmarking of several methods
}
\examples{
# This is the core function for running benchmarks
# Basic usage:
\dontrun{
result <- run_lri_methods(eset, design, dbs, methods = c("fgsea", "pca_limma"))
}
}
